+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# master
+
+- [Change] `.clickOnText` now matches current element as well
+
 # v1.0.0-alpha.1
 
 - [Feature] Can import properties directly `import { create } from 'path/to/page-object';`